Crate fatcat_api_spec

Re-exports

pub use self::client::Client;

pub use self::server::router;

Modules

client

models

router
Router provides fast and flexible routing for Iron.

server

Structs

ApiError
Very simple error type - just holds a description of the error. This is useful for human diagnosis and troubleshooting, but not for applications to parse. The justification for this is to deny applications visibility into the communication layer, forcing the application code to act solely on the logical responses that the API provides, promoting abstraction in the application code.

Context
Request context, both as received in a server handler or as sent in a client request. When REST microservices are chained, the Context passes data from the server API to any further HTTP requests.

ContextWrapper
Context wrapper, to bind an API with a context.

Enums

AcceptEditgroupResponse

CreateContainerBatchResponse

CreateContainerResponse

CreateCreatorBatchResponse

CreateCreatorResponse

CreateEditgroupResponse

CreateFileBatchResponse

CreateFileResponse

CreateReleaseBatchResponse

CreateReleaseResponse

CreateWorkBatchResponse

CreateWorkResponse

DeleteContainerResponse

DeleteCreatorResponse

DeleteFileResponse

DeleteReleaseResponse

DeleteWorkResponse

GetChangelogEntryResponse

GetChangelogResponse

GetContainerHistoryResponse

GetContainerResponse

GetCreatorHistoryResponse

GetCreatorReleasesResponse

GetCreatorResponse

GetEditgroupResponse

GetEditorChangelogResponse

GetEditorResponse

GetFileHistoryResponse

GetFileResponse

GetReleaseFilesResponse

GetReleaseHistoryResponse

GetReleaseResponse

GetStatsResponse

GetWorkHistoryResponse

GetWorkReleasesResponse

GetWorkResponse

LookupContainerResponse

LookupCreatorResponse

LookupFileResponse

LookupReleaseResponse

UpdateContainerResponse

UpdateCreatorResponse

UpdateFileResponse

UpdateReleaseResponse

UpdateWorkResponse

Traits

Api
API

ApiNoContext
API without a Context

ContextWrapperExt
Trait to extend an API to make it easy to bind it to a context.

Future
Trait for types which are a placeholder of a value that may become available at some later point in time.