Struct fatcat_api_spec::client::Client

pub struct Client { /* fields omitted */ }

A client that implements the API by making HTTP calls out to a server.

Methods

impl Client

pub fn try_new_http<T>(base_path: T) -> Result<Client, ClientInitError> where
T: IntoUrl,

pub fn try_new_https<T, CA>(
base_path: T,
ca_certificate: CA
) -> Result<Client, ClientInitError> where
T: IntoUrl,
CA: AsRef<Path>,

pub fn try_new_https_mutual<T, CA, K, C>(
base_path: T,
ca_certificate: CA,
client_key: K,
client_certificate: C
) -> Result<Client, ClientInitError> where
T: IntoUrl,
CA: AsRef<Path>,
K: AsRef<Path>,
C: AsRef<Path>,

pub fn try_new_with_hyper_client<T>(
base_path: T,
hyper_client: Arc<Fn() -> Client + Sync + Send>
) -> Result<Client, ClientInitError> where
T: IntoUrl,

Constructor for creating a Client by passing in a pre-made hyper client.

One should avoid relying on this function if possible, since it adds a dependency on the underlying transport implementation, which it would be better to abstract away. Therefore, using this function may lead to a loss of code generality, which may make it harder to move the application to a serverless environment, for example.

The reason for this function's existence is to support legacy test code, which did mocking at the hyper layer. This is not a recommended way to write new tests. If other reasons are found for using this function, they should be mentioned here.