Function fatcat_api_spec::server::route

pub fn route<T>(router: &mut Router, api: T) where
    T: Api + Send + Sync + Clone + 'static, 

Deprecated
: APIs should not overlap - only for use in legacy code.

Add routes for Api to a provided router.

Note that these routes are added straight onto the router. This means that if the router already has a route for an endpoint which clashes with those provided by this API, then the old route will be lost.

It is generally a bad idea to add routes in this way to an existing router, which may have routes on it for other APIs. Distinct APIs should be behind distinct paths to encourage separation of interfaces, which this function does not enforce. APIs should not overlap.

Alternative approaches include:

generate an iron::middleware::Handler (usually a router::Router or iron::middleware::chain) for each interface, and add those handlers inside an existing router, mounted at different paths - so the interfaces are separated by path

use a different instance of iron::Iron for each interface - so the interfaces are separated by the address/port they listen on

This function exists to allow legacy code, which doesn't separate its APIs properly, to make use of this crate.