Struct fatcat_api_spec::server::Authorization

pub struct Authorization {
    pub subject: String,
    pub scopes: Scopes,
    pub issuer: Option<String>,
}

Storage of authorization parameters for an incoming request, used for REST API authorization.

Fields
subject: String

Subject for which authorization is granted (i.e., what may be accessed.)

scopes: Scopes

Scopes for which authorization is granted (i.e., what types of access are permitted).

issuer: Option<String>

Identity of the party to whom authorization was granted, if available (i.e., who is responsible for the access).

In an OAuth environment, this is the identity of the client which issued an authorization request to the resource owner (end-user), and which has been directly authorized by the resource owner to access the protected resource. If the client delegates that authorization to another service (e.g., a proxy or other delegate), the issuer is still the original client which was authorized by the resource owner.